Fixed k<->E transformation, minor notation changes
Minor notation clarification in eq. 7 of the introductory text.
Fixed eq. 5 to agree with eq. 1 and code below.

Changed notation to disambiguate conflicts with formulas in the introductory text:
	energy_levels, En ("Delta" in text) -> channel_excitation_energies, channelExcitationEnergy
	delta ("En" in text) -> E_channel (channel kinetic energy)

Now "mass" is read in the code as the reduced mass of the system (and not twice as much).
Simplified energy <-> wave-number conversions and fixed missing hbarc.